♻️ Refactor measurement plaquettes
